Updated client now listens for acknowledgements from server when sending parameter metadata. Improved synchronization.
This better adheres to the defined protocol where every message is always met with a confirmation of that message. After the client sends the size of the individual members in the parameter-to-be-sent's data buffer, it should receive an OK before sending the number of member's in the parameter-to-be-sent's data buffer.