Modal windows and build folder removal
I set all windows to modal, so that only 1 of each window could be open at a time. This could turn out to be an issue if we had multiple of the same window open.
Also deleted the build folder, which has tens of thousands of lines of code, which is just for QtCreator, should not need for the client besides the .cpp, .h, and .ui files.

Oh and also added Answer option, which we wanted to have comments different from answers and how answers would be for only verified users and comments for anyone.