Updated documentation and added planning directory
The master-documentation.html file was updated to make use of the new documentation files added during sprint 2. Any markdown documentation files were transpiled into HTML files to play nicely with master-documentation.html. Notably, the new documentation is for the DALLO protocol, emailing, the cloud infrastructure, and the secure connection library. The planning directory includes to-dos and scratch notes from discussions surrounding plans to integrate the server and database, and how to add security (security is now a complete task).