Small Changes
Added Favorite Communities Tab.
Added more items to the list in Communities Window .
Added an invisible label in Profile that contains the file path of the profile picture, so the user can see the image when they click on it, this can be done away with easily.
Changed most mentions of "post" to "question."